Move metrics to package
